+++ syndicate/connection/api_gateway_connection.py
@@ -221,8 +221,7 @@ class ApiGatewayConnection(object):
         if request_models:
             params['requestModels'] = request_models
         if request_validator:
-            params['requestValidatorId'] = self.create_request_validator(
-                api_id, request_validator)
+            params['requestValidatorId'] = request_validator
         self.client.put_method(**params)
 
     def create_request_validator(self, api_id, request_validator):
@@ -649,3 +648,33 @@ class ApiGatewayConnection(object):
         params['patchOperations'] = [patchOperation, ]
         _LOG.debug(f'Updating rest api with params: "{params}"')
         return self.client.update_rest_api(**params)
+
+    def create_model(self, rest_api_id, name, content_type, description=None,
+                     schema=None):
+        """Adds a new Model resource to an existing RestApi resource."""
+        _LOG.debug(f'Creating new model "{name}"')
+        params = {
+            'contentType': content_type,
+            'restApiId': rest_api_id,
+            'name': name
+        }
+        if description:
+            params['description'] = description
+        if schema:
+            params['schema'] = schema
+        return self.client.create_model(**params)
+
+    def delete_model(self, rest_api_id, name):
+        _LOG.debug(f'Deleting model "{name}"')
+        return self.client.delete_model(restApiId=rest_api_id, modelName=name)
+
+    def get_model(self, rest_api_id, name, flatten=False):
+        try:
+            return self.client.get_model(restApiId=rest_api_id, modelName=name,
+                                         flatten=flatten)
+        except ClientError as e:
+            if 'NotFoundException' in str(e):
+                _LOG.warn(f'Cannot find model "{name}"')
+                return None
+            else:
+                raise e