+++ syndicate/core/resources/api_gateway_resource.py
@@ -13,6 +13,7 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 """
+import json
 import time
 from hashlib import md5
 
@@ -35,6 +36,16 @@ _CUSTOM_AUTHORIZER_TYPE = 'CUSTOM'
 
 POLICY_STATEMENT_SINGLETON = 'policy_statement_singleton'
 
+_REQUEST_VALIDATORS = {
+    'Validate body': {
+        'params': {'validate_request_body': True}, 'id': None},
+    'Validate query string parameters and headers': {
+        'params': {'validate_request_parameters': True}, 'id': None},
+    'Validate body, query string parameters, and headers': {
+        'params': {'validate_request_body': True,
+                   'validate_request_parameters': True}, 'id': None}
+}
+
 
 class ApiGatewayResource(BaseResource):
 
@@ -45,6 +56,29 @@ class ApiGatewayResource(BaseResource):
         self.cognito_res = cognito_res
         self.account_id = account_id
         self.region = region
+
+    def _create_default_validators(self, api_id):
+        for validator in _REQUEST_VALIDATORS.values():
+            params = validator['params']
+            _id = self.connection.create_request_validator(api_id, params)
+            validator['id'] = _id
+
+    def _retrieve_request_validator_id(self, api_id, request_validator=None):
+        if not request_validator:
+            return None
+        if request_validator.get('name'):
+            _id = self.connection.create_request_validator(api_id,
+                                                           request_validator)
+            return _id
+        if ('validate_request_body', True) in request_validator.items():
+            return _REQUEST_VALIDATORS['Validate body']['id']
+        elif ('validate_request_parameters', True) in request_validator\
+                .items():
+            return _REQUEST_VALIDATORS[
+                'Validate query string parameters and headers']['id']
+        else:
+            return _REQUEST_VALIDATORS[
+                'Validate body, query string parameters, and headers']['id']
 
     def api_resource_identifier(self, name, output=None):
         if output:
@@ -211,6 +245,9 @@ class ApiGatewayResource(BaseResource):
             binary_media_types=meta.get('binary_media_types'))
         api_id = api_item['id']
 
+        # create default request validators
+        self._create_default_validators(api_id)
+
         # set minimumCompressionSize if the param exists
         minimum_compression_size = meta.get('minimum_compression_size', None)
         if not minimum_compression_size:
@@ -253,6 +290,11 @@ class ApiGatewayResource(BaseResource):
                                                   'identity_source'),
                                               ttl=val.get('ttl'),
                                               provider_arns=provider_arns)
+
+        models = meta.get('models')
+        if models:
+            args = [{'api_id': api_id, 'models': {k: v}} for k, v in models.items()]
+            self.create_pool(self._create_model_from_metadata, args, 1)
         if api_resources:
             api_resp = meta.get('api_method_responses')
             api_integration_resp = meta.get('api_method_integration_responses')
@@ -499,14 +541,23 @@ class ApiGatewayResource(BaseResource):
             else:
                 authorization_type = _CUSTOM_AUTHORIZER_TYPE
 
+        method_request_models = method_meta.get('method_request_models')
+        if method_request_models:
+            (content_type, name), = method_request_models.items()
+            model = self.connection.get_model(api_id, name)
+            method_request_models = model if not model else method_request_models
+
+        request_validator_id = self._retrieve_request_validator_id(
+            api_id, method_meta.get('request_validator'))
+
         self.connection.create_method(
             api_id, resource_id, method,
             authorization_type=authorization_type,
             authorizer_id=method_meta.get('authorizer_id'),
             api_key_required=method_meta.get('api_key_required'),
             request_parameters=method_meta.get('method_request_parameters'),
-            request_models=method_meta.get('method_request_models'),
-            request_validator=method_meta.get('request_validator'))
+            request_models=method_request_models,
+            request_validator=request_validator_id)
         # second step: create integration
         integration_type = method_meta.get('integration_type')
         # set up integration - lambda or aws service
@@ -684,3 +735,15 @@ class ApiGatewayResource(BaseResource):
                 _LOG.warn('API Gateway %s is not found', api_id)
             else:
                 raise e
+
+    @unpack_kwargs
+    def _create_model_from_metadata(self, api_id, models):
+        _LOG.info('Going to process API Gateway models')
+        for name, model_data in models.items():
+            description = model_data.get('description')
+            schema = model_data.get('schema')
+            if isinstance(schema, dict):
+                schema = json.dumps(schema)
+            content_type = model_data.get('content_type')
+            self.connection.create_model(
+                api_id, name, content_type, description, schema)