+++ CHANGELOG.md
@@ -29,10 +29,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed an issue when build meta is validated before resolving aliases in it
 - Added `minimum_compression_size` param to API gateway deployment_resource declaration. It enables payload compression for an API with given size;
 - Added lambda URL support;
-- Add lambda ephemeral storage support;
+- Added lambda ephemeral storage support;
 - Added an ability to create DynamoDB in on-demand mode if `read_capacity` 
   and `write_capacity` are not specified;
 - Describe an existing Api Gateway instead of creating a new one;
+- Added an ability to create models in API Gateway;
 
 
 ## [0.9.6] - 2021-04-29